requirements.txt
fastapi
uvicorn
requests
streamlit
python-dotenv
transformers
torch
weaviate-client
sentence-transformers